requirements.txt
# Dépendances principales
numpy>=1.26.4
librosa>=0.10.1
soundfile>=0.13.0
scipy>=1.12.0
python-dotenv>=1.0.0

# Transcription audio
faster-whisper>=0.10.0
torch>=2.2.0
torchaudio>=2.2.0
torchvision>=0.17.0

# Diarisation
pyannote.audio>=3.3.2
pytorch-metric-learning>=2.8.1
speechbrain>=1.0.3
transformers>=4.38.2

# Traitement et analyse
pandas>=2.2.2
matplotlib>=3.9.1
scikit-learn>=1.6.1

# Utilitaires
tqdm>=4.66.4
rich>=13.9.4
typer>=0.12.5

# Core Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
scikit-learn==1.3.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
noisereduce==2.0.1

# Deep Learning
torch==2.1.0
torchaudio==2.1.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2

# File Handling and I/O
openpyxl==3.1.2
python-dotenv==1.0.0

# Utilities and Testing
tqdm==4.65.0
pytest==7.4.0

# Scientific and Machine Learning
scikit-learn==1.3.0

# Débruitage avancé basé sur deep learning (optionnel mais recommandé)
# Décommentez la ligne suivante pour installer DeepFilterNet
deepfilternet>=0.5.0

# Dépendances pour la détection d'activité vocale (VAD)
# Le modèle Silero VAD est téléchargé automatiquement via torch.hub